requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
anyio==4.5.2
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
azure-core==1.31.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blis==1.1.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
diffprivlib==0.6.5
dill==0.3.8
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
exceptiongroup==1.2.2
executing==2.1.0
Faker==30.4.0
fastapi==0.115.2
fastjsonschema==2.21.1
fasttext-wheel==0.9.2
filelock==3.16.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.24.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pexpect==4.9.0
phonenumbers==8.13.47
pillow==11.0.0
pipdeptree==2.24.0
platformdirs==4.3.6
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.18
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge_score==0.1.2
rpds-py==0.22.3
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.3
sentence-transformers==1.1.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-alignments==0.9.1
spacy-curated-transformers==0.3.0
spacy-huggingface-pipelines==0.0.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
spacy_fastlang==2.1.0
srsly==2.4.8
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.3
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tinycss2==1.4.0
tldextract==5.1.2
tokenizers==0.13.3
tomli==2.2.1
torch==2.2.2
torch-vision==0.1.6.dev0
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.66.5
tr_core_news_trf @ file:./tr_core_news_trf-3.4.2-py3-none-any.whl
traitlets==5.14.3
transformers==4.28.0
typer==0.7.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
